+++ dependencies/ominipg.ts
@@ -1,1 +1,1 @@
-export * from "jsr:@oxian/ominipg@0.9.0-rc.3";
+export * from "jsr:@oxian/ominipg@0.9.0-rc.5";