+++ dependencies/oxian-hypervisor.ts
@@ -1,1 +1,1 @@
-export * from "jsr:@oxian/oxian-js@0.20.0-rc.7/hypervisor";
+export * from "jsr:@oxian/oxian-js@0.21.0-rc.2/hypervisor";