+++ dependencies/oxian-worker.ts
@@ -1,1 +1,1 @@
-export * from "jsr:@oxian/oxian-js@0.20.0-rc.7/worker";
+export * from "jsr:@oxian/oxian-js@0.21.0-rc.2/worker";